mtb_1/home/ubuntu/solana_trading_bot/requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
selenium==4.10.0
webdriver-manager==3.8.6
solana==0.30.2
pyserum==0.5.0
jupag-py==0.1.0
aiohttp==3.8.5
asyncio==3.4.3
python-telegram-bot==13.15
loguru==0.7.0
pandas==2.0.3
numpy==1.24.3
